moved stream class into odm
